requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
django-filter==23.2
factory-boy==3.3.0
